Add UploadedFile base class and Spreadsheet impl
We need protection around uploaded spreadsheet that are being submitted
for addition to a data set.

This needs to check whether the file looks right:
  - Is not too big or too small,
  - Has a valid content-type,
  - Not a virus!

I've pulled in the csv and excel parsing code from Backdrop and tweaked
them to fit the new admin app. This was done to retain a certain amount
of parity between the new and old uploader.