Set redis session expiry time to 2 hours
This is to match signon expiry time to minimise the cases where a signon
user session has expired, but the admin app user session has not
expired. This leads to 403s when we request data from stagecraft, which
will be caught, but the session length being a similar time will avoid
this in most cases.