Add assertion that session doesn't contain element
Add a method assert_session_does_not_contain that asserts that a
KeyError is raised for a given key when accessing the session.

Change the existing assert_session_contains method so that it
is the only thing checking for a KeyError and raising an
AssertionError because of that. This is because otherwise
we would be asserting that an AssertionError is raised, which
is confusing.