Upgrade performanceplatform-client
The new endpoints in the admin API are now supported in the admin client
(AdminAPI). Make use of these in the admin app to get the benefit of
things like backoff and gzipping, and simplify the admin app code
slightly.

Controversially, this introduces a TESTING environment variable that is
set during testing, and used in production code. This is because the
admin app needs to speak to the admin API at startup to get the
available module types for the form select field.