Delete dashboard session on edit form
The 'pending_dashboard' session variable is used for passing saved state
from the create / update routes back to the form rendering routes. As
such it should only live for that single request. If the pending
dashboard stays in the session beyond that we risk showing unexpected
(and arguably incorrect) data in the form in certain situations. For
example, if a user goes back to the list page before saving and then
clicks back into the same dashboard the data shown would not be the data
that's stored in stagecraft but stale data from the session.