Add WTForms field that pretty-prints JSON
When editing JSON in the admin app, it gets difficult to read a large
chunk of it.

This commit pretty-prints JSON by adding a new WTForms field type
that reads the JSON string in using json.loads() and then calls
json.dumps() with the indent and sort_keys options.

We should think about whether it would be useful to commit this
upstream.