Implement support for signonotron callbacks
When a user is suspended in signon, client apps can receive
notifications of these events.

We should handle these, so that people are logged out as appropriate.

This change handles the session state for the UI, which is what we care
about in this instance. There is a corresponding change in pp-puppet to
ensure that the request is handled by this application, rather than
being proxied through to stagecraft.

This fixes the bug where only clearing the session in stagecraft wasn’t
logging actual users out of the admin UI.