Upload data for the digital take-up module.
Create a digital take up module if it doesn't exist.
Create a data group, data set and transform if they don't exist

Reuses the existing upload spreadsheet functionality.

Display website to the user in the checkbox channel choices,
but the spreadsheet contains a channel named digital as this is the
current way data is uploaded to backdrop

Replace the add digital take-up button with a success message if a
digital take-up module already exists for the dashboard.

Protect the digital take-up upload routes so that they redirect
users back to the dashboard hub page if the digital take-up
module exists.