wip - working tests for reorder on create
should_use_session will always return true for create which seems
unecessary if the data is always the same as in the form. We should
think about changing this however it still needs the form data identity
check for the tests as if the form is overwritten with the session data
then validations fail despite the data being identical (and valid). This
is weird and should be looked into.