Start on clone dashboard
This enables a link of the format 'admin/dashboards/new?clone_module={module_id}&modules=1' or 'admin/dashboards/dashboard_uuid?clone_module={module_id}&modules=1' which when followed will redirect to a new or the existing dashboard with anything in the session still there and the new module prefilled and present.

Currently the links are available in the sidebar but this will need to
move to a popover as the sidebar is not big enough for this to work
well.

Working unstyled module

For some reason I can't get the form post way to work. It clears the
dashboard details. This also clears any sections or modules because the
session is cleared?

Better clone module - keeps section and empty module