Enable clone dashboard
When the user clicks 'Clone module' on a new or existing dashboard it
will redirect to a page containing a dropdown of dashboards to select
from and the form progress will be saved in the session (because of the
update_modules_form_and_redirect decorator). When the user selects a
dashboard and submits it will redirect back to the page, but now there
will also be a dropdown containing modules on the selected dashboard.
Clicking a link in this dropdown will redirect to the original dashboard
page with the module now appended to the form using the following
mechanism:

A link of the format 'admin/dashboards/new?clone_module={module_id}' or 'admin/dashboards/dashboard_uuid?clone_module={module_id}' which when followed will redirect to a new or the existing dashboard with anything in the session still there and the new module prefilled and present.