UI improvements
Tidy up the styling for long dashboard titles.

Add in more descriptive links in dropdown.

Make clear you are selecting a source dashboard for modules.

Provide a back link to the dashboard.

Tell the user what dashboard you are cloning into.

Aria labels should be more consistent, styling should be a little less
haphazard, it should be responsive. This is an internal tool though so
perhaps we can leave for now.