Fix tests around calculating starts of quarters
python doesn't understand negative or 0 months - it doesn't go back
to the previous year for instance. Use a range to get the month. Use dtend to prevent the wrong
number of quarters, don't just cut off the end.

There may be a better way to do this but I struggled with it for a
while.