Add Cache-Control header to all requests
The lack of a Cache-Control header makes some upstream hosts assume
that requests to this app can be cached.

Add a Cache-Control header for all requests so that they aren't
cache by things upstream from us.

This should be improved in the future so that requests for cacheable
resources are actually cached.