Ensure hub renders when modules have no datatype
The hub page failed to render when modules have no associated datatype
because it assumes the key is always there. There are modules such as
tabs that will not have an associated datatype as this is contained
within the json config for the tabs.

Also fix some pep8 errors picked up by autopep8