Display error page if datasets list fail to load.
When trying to access the upload data page in the self serve app,
if an http error is returned from stagecraft, then this exception
is just thrown back to the user, e.g. a 500 Internal Server Error
is returned and not a nice validation message.

The only exception to this are 401s. They redirect the user to a
sign out page.

Changed this behaviour to display an error page if any http error
except a 401 is returned from stagecraft.

This uses Flask's errorhandler as suggested by Tim. Handling more
errors in this way is desirable, but a refactor is beyond the scope
of this story.