Check for dates when parsing uploaded csv files.
When data is uploaded in a csv file, parse_csv only formated the
data in the cell if it was a number. Dates and strings were posted
to backdrop unchanged.

If users uploaded data for the same date, but the timestamp was
formatted differently in each row, another record would be stored
in that date.

i.e. The following would create two separate rows in backdrop:
2015-10-05T00:00:00Z,week,1
2015-10-05T00:00:00,week,100

Updated parse_csv to also check whether the cell can be parsed as
a date and then convert the date to utc format.