Check whether text cells contain dates when parsing uploaded xls files.
When data is uploaded in an excel file, parse_excel only converts data
in a cell of type date to a utc date. If the date is a string,
parse_excel does not format it, and posts the date to backdrop unchanged.

If users uploaded data for the same date, but the timestamp was
formatted differently in each row, e.g if one row was of type date and
another text, but the data in them was the same, two records would be
stored for that date in backdrop.

Updated parse_excel to also check whether a cell of type text can be
parsed as a date and then convert the date to utc format.